漢彬-12/29 article的發文時間修正完畢

+++ src/main/java/foodelicious/article/controller/ArticleController.java
@@ -40,7 +40,7 @@ public class ArticleController {
 		System.out.println(params);
 		System.out.println(params.get("article"));
 		System.out.println(params.get("classify"));
-		int id = (int) session.getAttribute("userID");
+		Long id = (Long) session.getAttribute("userID");
 		
 		System.out.println(id);
 		articleService.pushArticle(params, id);
